feat: changing how global middlewares work
