+++ neko-router/src/MiddlewareFactory.ts
@@ -1,8 +1,8 @@
 import { Middleware } from './Middleware';
-import { LexerToken, Request, Response } from './types';
+import { HandlerType, LexerToken, Request, Response } from './types';
 
 export class MiddlewareFactory {
-  public static create(func: Function): Middleware {
+  public static create(func: HandlerType): Middleware {
     const cls = class extends Middleware {
       call(req: Request, res: Response, next: () => Promise<void>): Promise<void> {
         return func(req, res, next);
